[DURACOM-271] Renamed and reorganized suggestion data services
