+++ src/app/core/notifications/suggestions/models/suggestion-objects.resource-type.ts
@@ -1,4 +1,4 @@
-import { ResourceType } from '../../shared/resource-type';
+import { ResourceType } from '../../../shared/resource-type';
 
 /**
  * The resource type for the Suggestion object
