+++ src/app/core/notifications/suggestions/models/suggestion-source-object.resource-type.ts
@@ -1,4 +1,4 @@
-import { ResourceType } from '../../shared/resource-type';
+import { ResourceType } from '../../../shared/resource-type';
 
 /**
  * The resource type for the Suggestion Source object
