+++ src/app/core/notifications/suggestions/models/suggestion-source.model.ts
@@ -3,11 +3,11 @@ import {
   deserialize,
 } from 'cerialize';
 
-import { typedObject } from '../../cache/builders/build-decorators';
-import { CacheableObject } from '../../cache/cacheable-object.model';
-import { HALLink } from '../../shared/hal-link.model';
-import { ResourceType } from '../../shared/resource-type';
-import { excludeFromEquals } from '../../utilities/equals.decorators';
+import { typedObject } from '../../../cache/builders/build-decorators';
+import { CacheableObject } from '../../../cache/cacheable-object.model';
+import { HALLink } from '../../../shared/hal-link.model';
+import { ResourceType } from '../../../shared/resource-type';
+import { excludeFromEquals } from '../../../utilities/equals.decorators';
 import { SUGGESTION_SOURCE } from './suggestion-source-object.resource-type';
 
 /**
