+++ src/app/core/notifications/suggestions/source/suggestion-source-data.service.ts
@@ -3,22 +3,22 @@ import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
-import { NotificationsService } from '../../../shared/notifications/notifications.service';
-import { FollowLinkConfig } from '../../../shared/utils/follow-link-config.model';
-import { RemoteDataBuildService } from '../../cache/builders/remote-data-build.service';
-import { ObjectCacheService } from '../../cache/object-cache.service';
-import { CoreState } from '../../core-state.model';
+import { NotificationsService } from '../../../../shared/notifications/notifications.service';
+import { FollowLinkConfig } from '../../../../shared/utils/follow-link-config.model';
+import { RemoteDataBuildService } from '../../../cache/builders/remote-data-build.service';
+import { ObjectCacheService } from '../../../cache/object-cache.service';
+import { CoreState } from '../../../core-state.model';
 import {
   FindAllData,
   FindAllDataImpl,
-} from '../../data/base/find-all-data';
-import { IdentifiableDataService } from '../../data/base/identifiable-data.service';
-import { DefaultChangeAnalyzer } from '../../data/default-change-analyzer.service';
-import { FindListOptions } from '../../data/find-list-options.model';
-import { PaginatedList } from '../../data/paginated-list.model';
-import { RemoteData } from '../../data/remote-data';
-import { RequestService } from '../../data/request.service';
-import { HALEndpointService } from '../../shared/hal-endpoint.service';
+} from '../../../data/base/find-all-data';
+import { IdentifiableDataService } from '../../../data/base/identifiable-data.service';
+import { DefaultChangeAnalyzer } from '../../../data/default-change-analyzer.service';
+import { FindListOptions } from '../../../data/find-list-options.model';
+import { PaginatedList } from '../../../data/paginated-list.model';
+import { RemoteData } from '../../../data/remote-data';
+import { RequestService } from '../../../data/request.service';
+import { HALEndpointService } from '../../../shared/hal-endpoint.service';
 import { SuggestionSource } from '../models/suggestion-source.model';
 
 /**
