+++ src/app/core/notifications/suggestions/suggestion-data.service.spec.ts
@@ -1,5 +1,4 @@
 import { HttpClient } from '@angular/common/http';
-import { Store } from '@ngrx/store';
 import {
   cold,
   getTestScheduler,
@@ -10,51 +9,44 @@ import { TestScheduler } from 'rxjs/testing';
 import { NotificationsService } from '../../../shared/notifications/notifications.service';
 import { createSuccessfulRemoteDataObject$ } from '../../../shared/remote-data.utils';
 import { RemoteDataBuildService } from '../../cache/builders/remote-data-build.service';
+import { RequestParam } from '../../cache/models/request-param.model';
 import { ObjectCacheService } from '../../cache/object-cache.service';
 import { RestResponse } from '../../cache/response.models';
-import { CoreState } from '../../core-state.model';
-import { FindAllData } from '../../data/base/find-all-data';
-import { testFindAllDataImplementation } from '../../data/base/find-all-data.spec';
-import { DefaultChangeAnalyzer } from '../../data/default-change-analyzer.service';
 import { RemoteData } from '../../data/remote-data';
-import { GetRequest } from '../../data/request.models';
 import { RequestService } from '../../data/request.service';
 import { RequestEntry } from '../../data/request-entry.model';
 import { RequestEntryState } from '../../data/request-entry-state.model';
 import { HALEndpointService } from '../../shared/hal-endpoint.service';
-import { SuggestionSource } from '../models/suggestion-source.model';
-import { SuggestionSourceDataService } from './suggestion-source-data.service';
+import { SuggestionDataService } from './suggestion-data.service';
 
-describe('SuggestionSourceDataService test', () => {
+describe('SuggestionDataService test', () => {
   let scheduler: TestScheduler;
-  let service: SuggestionSourceDataService;
+  let service: SuggestionDataService;
   let requestService: RequestService;
   let rdbService: RemoteDataBuildService;
   let objectCache: ObjectCacheService;
   let halService: HALEndpointService;
   let notificationsService: NotificationsService;
   let http: HttpClient;
-  let comparator: DefaultChangeAnalyzer<SuggestionSource>;
   let responseCacheEntry: RequestEntry;
 
-  const store = {} as Store<CoreState>;
-  const endpointURL = `https://rest.api/rest/api/suggestionsources`;
-  const requestUUID = '8b3c613a-5a4b-438b-9686-be1d5b4a1c5a';
 
+  const testSource = 'test-source';
+  const testUserId = '1234-4321';
+  const endpointURL = `https://rest.api/rest/api/`;
+  const requestUUID = '8b3c613a-5a4b-438b-9686-be1d5b4a1c5a';
   const remoteDataMocks = {
     Success: new RemoteData(null, null, null, RequestEntryState.Success, null, null, 200),
   };
 
   function initTestService() {
-    return new SuggestionSourceDataService(
+    return new SuggestionDataService(
       requestService,
       rdbService,
-      store,
       objectCache,
       halService,
       notificationsService,
       http,
-      comparator,
     );
   }
 
@@ -64,7 +56,6 @@ describe('SuggestionSourceDataService test', () => {
     objectCache = {} as ObjectCacheService;
     http = {} as HttpClient;
     notificationsService = {} as NotificationsService;
-    comparator = {} as DefaultChangeAnalyzer<SuggestionSource>;
     responseCacheEntry = new RequestEntry();
     responseCacheEntry.request = { href: 'https://rest.api/' } as any;
     responseCacheEntry.response = new RestResponse(true, 200, 'Success');
@@ -75,6 +66,7 @@ describe('SuggestionSourceDataService test', () => {
       removeByHrefSubstring: {},
       getByHref: observableOf(responseCacheEntry),
       getByUUID: observableOf(responseCacheEntry),
+      setStaleByHrefSubstring: observableOf(true),
     });
 
     halService = jasmine.createSpyObj('halService', {
@@ -86,32 +78,30 @@ describe('SuggestionSourceDataService test', () => {
       buildList: cold('a', { a: remoteDataMocks.Success }),
     });
 
-
     service = initTestService();
   });
 
-  describe('composition', () => {
-    const initFindAllService = () => new SuggestionSourceDataService(null, null, null, null, null, null, null, null) as unknown as FindAllData<any>;
-    testFindAllDataImplementation(initFindAllService);
-  });
+  describe('Suggestion service', () => {
+    it('should call suggestionsDataService.searchBy', () => {
+      spyOn((service as any).searchData, 'searchBy').and.returnValue(observableOf(null));
+      const options = {
+        searchParams: [new RequestParam('target', testUserId), new RequestParam('source', testSource)],
+      };
+      service.getSuggestionsByTargetAndSource(testUserId, testSource);
+      expect((service as any).searchData.searchBy).toHaveBeenCalledWith('findByTargetAndSource', options, false, true);
+    });
 
-  describe('getSources', () => {
-    it('should send a new GetRequest', () => {
-      const expected = new GetRequest(requestService.generateRequestId(),  `${endpointURL}`);
-      scheduler.schedule(() => service.getSources().subscribe());
-      scheduler.flush();
-
-      expect(requestService.send).toHaveBeenCalledWith(expected, true);
+    it('should call suggestionsDataService.delete', () => {
+      spyOn((service as any).deleteData, 'delete').and.returnValue(observableOf(null));
+      service.deleteSuggestion('1');
+      expect((service as any).deleteData.delete).toHaveBeenCalledWith('1');
     });
   });
 
-  describe('getSource', () => {
-    it('should send a new GetRequest', () => {
-      const expected = new GetRequest(requestService.generateRequestId(),  `${endpointURL}/testId`);
-      scheduler.schedule(() => service.getSource('testId').subscribe());
-      scheduler.flush();
-
-      expect(requestService.send).toHaveBeenCalledWith(expected, true);
+  describe('Request service', () => {
+    it('should call requestService.setStaleByHrefSubstring', () => {
+      service.clearSuggestionRequests();
+      expect(requestService.setStaleByHrefSubstring).toHaveBeenCalled();
     });
   });
 });
