+++ src/app/core/provide-core.ts
@@ -30,11 +30,11 @@ import { EPerson } from './eperson/models/eperson.model';
 import { Group } from './eperson/models/group.model';
 import { MetadataField } from './metadata/metadata-field.model';
 import { MetadataSchema } from './metadata/metadata-schema.model';
-import { SuggestionSource } from './notifications/models/suggestion-source.model';
-import { SuggestionTarget } from './notifications/models/suggestion-target.model';
 import { QualityAssuranceEventObject } from './notifications/qa/models/quality-assurance-event.model';
 import { QualityAssuranceSourceObject } from './notifications/qa/models/quality-assurance-source.model';
 import { QualityAssuranceTopicObject } from './notifications/qa/models/quality-assurance-topic.model';
+import { SuggestionSource } from './notifications/suggestions/models/suggestion-source.model';
+import { SuggestionTarget } from './notifications/suggestions/models/suggestion-target.model';
 import { OrcidHistory } from './orcid/model/orcid-history.model';
 import { OrcidQueue } from './orcid/model/orcid-queue.model';
 import { ResearcherProfile } from './profile/model/researcher-profile.model';
