+++ src/app/notifications/suggestion-list-element/suggestion-approve-and-import.ts
@@ -1,4 +1,4 @@
-import { Suggestion } from '../../core/notifications/models/suggestion.model';
+import { Suggestion } from '../../core/notifications/suggestions/models/suggestion.model';
 
 /**
  * A simple interface to unite a specific suggestion and the id of the chosen collection
