+++ src/app/notifications/suggestion-list-element/suggestion-evidences/suggestion-evidences.component.ts
@@ -8,7 +8,7 @@ import {
 } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 
-import { SuggestionEvidences } from '../../../core/notifications/models/suggestion.model';
+import { SuggestionEvidences } from '../../../core/notifications/suggestions/models/suggestion.model';
 import { fadeIn } from '../../../shared/animations/fade';
 import { ObjectKeysPipe } from '../../../shared/utils/object-keys-pipe';
