+++ src/app/notifications/suggestion-targets/suggestion-targets.actions.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-classes-per-file */
 import { Action } from '@ngrx/store';
 
-import { SuggestionTarget } from '../../core/notifications/models/suggestion-target.model';
+import { SuggestionTarget } from '../../core/notifications/suggestions/models/suggestion-target.model';
 import { type } from '../../shared/ngrx/type';
 
 /**
