+++ src/app/notifications/suggestion-targets/suggestion-targets.reducer.ts
@@ -1,4 +1,4 @@
-import { SuggestionTarget } from '../../core/notifications/models/suggestion-target.model';
+import { SuggestionTarget } from '../../core/notifications/suggestions/models/suggestion-target.model';
 import {
   SuggestionTargetActionTypes,
   SuggestionTargetsActions,
