+++ src/app/notifications/suggestion.service.spec.ts
@@ -7,9 +7,9 @@ import {
   SortOptions,
 } from '../core/cache/models/sort-options.model';
 import { FindListOptions } from '../core/data/find-list-options.model';
-import { SuggestionTarget } from '../core/notifications/models/suggestion-target.model';
-import { SuggestionsDataService } from '../core/notifications/suggestions-data.service';
-import { SuggestionTargetDataService } from '../core/notifications/target/suggestion-target-data.service';
+import { SuggestionTarget } from '../core/notifications/suggestions/models/suggestion-target.model';
+import { SuggestionDataService } from '../core/notifications/suggestions/suggestion-data.service';
+import { SuggestionTargetDataService } from '../core/notifications/suggestions/target/suggestion-target-data.service';
 import { ResearcherProfile } from '../core/profile/model/researcher-profile.model';
 import { ResearcherProfileDataService } from '../core/profile/researcher-profile-data.service';
 import { ResourceType } from '../core/shared/resource-type';
@@ -22,7 +22,7 @@ describe('SuggestionsService test', () => {
   let scheduler: TestScheduler;
   let service: SuggestionsService;
   let researcherProfileService: ResearcherProfileDataService;
-  let suggestionsDataService: SuggestionsDataService;
+  let suggestionsDataService: SuggestionDataService;
   let suggestionTargetDataService: SuggestionTargetDataService;
   let translateService: any = {
     instant: (str) => str,
@@ -89,7 +89,7 @@ describe('SuggestionsService test', () => {
         sort: sortOptions,
       };
       service.getTargets('source', 10, 1);
-      expect(suggestionTargetDataService.getTargets).toHaveBeenCalledWith('source', findListOptions);
+      expect(suggestionTargetDataService.getTargetsBySource).toHaveBeenCalledWith('source', findListOptions);
     });
 
     it('should get suggestions', () => {
