+++ src/app/shared/mocks/publication-claim.mock.ts
@@ -2,8 +2,8 @@
 // REST Mock ---------------------------------------------------------------------
 // -------------------------------------------------------------------------------
 
-import { Suggestion } from '../../core/notifications/models/suggestion.model';
-import { SUGGESTION } from '../../core/notifications/models/suggestion-objects.resource-type';
+import { Suggestion } from '../../core/notifications/suggestions/models/suggestion.model';
+import { SUGGESTION } from '../../core/notifications/suggestions/models/suggestion-objects.resource-type';
 
 export const mockSuggestionPublicationOne: Suggestion =  {
   id: '24694773',
