+++ src/app/suggestions-page/suggestions-page.component.ts
@@ -37,8 +37,8 @@ import {
 import { FindListOptions } from '../core/data/find-list-options.model';
 import { PaginatedList } from '../core/data/paginated-list.model';
 import { RemoteData } from '../core/data/remote-data';
-import { Suggestion } from '../core/notifications/models/suggestion.model';
-import { SuggestionTarget } from '../core/notifications/models/suggestion-target.model';
+import { Suggestion } from '../core/notifications/suggestions/models/suggestion.model';
+import { SuggestionTarget } from '../core/notifications/suggestions/models/suggestion-target.model';
 import { PaginationService } from '../core/pagination/pagination.service';
 import { redirectOn4xx } from '../core/shared/authorized.operators';
 import { getFirstSucceededRemoteDataPayload } from '../core/shared/operators';
